[DDW-683] Separated Connecting Syncing into smaller components
