+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnecting.js
@@ -1,17 +1,11 @@
 // @flow
 import React, { Component } from 'react';
-import SVGInline from 'react-svg-inline';
 import { observer } from 'mobx-react';
-import { defineMessages, intlShape } from 'react-intl';
 import classNames from 'classnames';
-import { Button } from 'react-polymorph/lib/components/Button';
-import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
 import StatusIcons from './StatusIcons';
-import LoadingSpinner from '../../widgets/LoadingSpinner';
-import daedalusLogo from '../../../assets/images/daedalus-logo-loading-grey.inline.svg';
-import linkNewWindow from '../../../assets/images/link-ic.inline.svg';
-import adaLogo from '../../../assets/images/ada-logo.inline.svg';
-import cardanoLogo from '../../../assets/images/cardano-logo.inline.svg';
+import ReportIssue from './ReportIssue';
+import LogosDisplay from './LogosDisplay';
+import SyncingConnectingTitle from './SyncingConnectingTitle';
 import { CardanoNodeStates } from '../../../../../common/types/cardano-node.types';
 import styles from './SyncingConnecting.scss';
 import type { CardanoNodeState } from '../../../../../common/types/cardano-node.types';
@@ -19,97 +13,6 @@ import { REPORT_ISSUE_TIME_TRIGGER } from '../../../config/timingConfig';
 
 let connectingInterval = null;
 let syncingInterval = null;
-
-const messages = defineMessages({
-  starting: {
-    id: 'loading.screen.startingCardanoMessage',
-    defaultMessage: '!!!Starting Cardano node',
-    description: 'Message "Starting Cardano node" on the loading screen.',
-  },
-  stopping: {
-    id: 'loading.screen.stoppingCardanoMessage',
-    defaultMessage: '!!!Stopping Cardano node',
-    description: 'Message "Stopping Cardano node" on the loading screen.',
-  },
-  stopped: {
-    id: 'loading.screen.stoppedCardanoMessage',
-    defaultMessage: '!!!Cardano node stopped',
-    description: 'Message "Cardano node stopped" on the loading screen.',
-  },
-  updating: {
-    id: 'loading.screen.updatingCardanoMessage',
-    defaultMessage: '!!!Updating Cardano node',
-    description: 'Message "Updating Cardano node" on the loading screen.',
-  },
-  updated: {
-    id: 'loading.screen.updatedCardanoMessage',
-    defaultMessage: '!!!Cardano node updated',
-    description: 'Message "Cardano node updated" on the loading screen.',
-  },
-  crashed: {
-    id: 'loading.screen.crashedCardanoMessage',
-    defaultMessage: '!!!Cardano node crashed',
-    description: 'Message "Cardano node crashed" on the loading screen.',
-  },
-  unrecoverable: {
-    id: 'loading.screen.unrecoverableCardanoMessage',
-    defaultMessage:
-      '!!!Unable to start Cardano node. Please submit a support request.',
-    description:
-      'Message "Unable to start Cardano node. Please submit a support request." on the loading screen.',
-  },
-  connecting: {
-    id: 'loading.screen.connectingToNetworkMessage',
-    defaultMessage: '!!!Connecting to network',
-    description: 'Message "Connecting to network" on the loading screen.',
-  },
-  reconnecting: {
-    id: 'loading.screen.reconnectingToNetworkMessage',
-    defaultMessage: '!!!Network connection lost - reconnecting',
-    description:
-      'Message "Network connection lost - reconnecting" on the loading screen.',
-  },
-  syncing: {
-    id: 'loading.screen.syncingBlocksMessage',
-    defaultMessage: '!!!Syncing blocks',
-    description: 'Message "Syncing blocks" on the loading screen.',
-  },
-  reportConnectingIssueText: {
-    id: 'loading.screen.reportIssue.connecting.text',
-    defaultMessage: '!!!Having trouble connecting to network?',
-    description: 'Report connecting issue text on the loading screen.',
-  },
-  reportSyncingIssueText: {
-    id: 'loading.screen.reportIssue.syncing.text',
-    defaultMessage: '!!!Having trouble syncing?',
-    description: 'Report syncing issue text on the loading screen.',
-  },
-  reportIssueButtonLabel: {
-    id: 'loading.screen.reportIssue.buttonLabel',
-    defaultMessage: '!!!Open support ticket',
-    description: 'Open support ticket button label on the loading.',
-  },
-  reportIssueDownloadLogsLinkLabel: {
-    id: 'loading.screen.reportIssue.downloadLogsLinkLabel',
-    defaultMessage: '!!!Download logs',
-    description: 'Download logs button label on the loading.',
-  },
-  tlsCertificateNotValidError: {
-    id: 'loading.screen.errors.tlsCertificateNotValidPleaseRestartError',
-    defaultMessage: '!!!TLS certificate is not valid, please restart Daedalus.',
-    description: 'The TLS cert is not valid and Daedalus should be restarted',
-  },
-  loadingWalletData: {
-    id: 'loading.screen.loadingWalletData',
-    defaultMessage: '!!!Loading wallet data',
-    description: 'Message "Loading wallet data" on the loading screen.',
-  },
-  reportIssueButtonUrl: {
-    id: 'loading.screen.reportIssue.reportIssueButtonUrl',
-    defaultMessage: '!!!https://iohk.zendesk.com/hc/en-us/requests/new/',
-    description: 'Link to Open Support page',
-  },
-});
 
 type State = {
   connectingTime: number,
@@ -122,6 +25,8 @@ type Props = {
   hasBeenConnected: boolean,
   isConnected: boolean,
   isSynced: boolean,
+  isConnecting: boolean,
+  isSyncing: boolean,
   isNodeStopping: boolean,
   isNodeStopped: boolean,
   isTlsCertInvalid: boolean,
@@ -144,10 +49,6 @@ type Props = {
 
 @observer
 export default class SyncingConnecting extends Component<Props, State> {
-  static contextTypes = {
-    intl: intlShape.isRequired,
-  };
-
   state = {
     connectingTime: 0,
     syncingTime: 0,
@@ -257,104 +158,13 @@ export default class SyncingConnecting extends Component<Props, State> {
     }
   };
 
-  _getConnectingMessage = () => {
-    const { cardanoNodeState, hasBeenConnected, isTlsCertInvalid } = this.props;
-    let connectingMessage;
-    switch (cardanoNodeState) {
-      case null:
-      case CardanoNodeStates.STARTING:
-        connectingMessage = messages.starting;
-        break;
-      case CardanoNodeStates.STOPPING:
-      case CardanoNodeStates.EXITING:
-        connectingMessage = messages.stopping;
-        break;
-      case CardanoNodeStates.STOPPED:
-        connectingMessage = messages.stopped;
-        break;
-      case CardanoNodeStates.UPDATING:
-        connectingMessage = messages.updating;
-        break;
-      case CardanoNodeStates.UPDATED:
-        connectingMessage = messages.updated;
-        break;
-      case CardanoNodeStates.CRASHED:
-      case CardanoNodeStates.ERRORED:
-        connectingMessage = messages.crashed;
-        break;
-      case CardanoNodeStates.UNRECOVERABLE:
-        connectingMessage = messages.unrecoverable;
-        break;
-      default:
-        // also covers CardanoNodeStates.RUNNING state
-        connectingMessage = hasBeenConnected
-          ? messages.reconnecting
-          : messages.connecting;
-    }
-    const isConnectingMessage =
-      connectingMessage === messages.connecting ||
-      connectingMessage === messages.reconnecting;
-    if (isTlsCertInvalid && isConnectingMessage) {
-      return messages.tlsCertificateNotValidError;
-    }
-    return connectingMessage;
-  };
-
-  _renderLoadingScreen = () => {
-    const { intl } = this.context;
-    const {
-      isConnected,
-      isSynced,
-      isNodeStopping,
-      isNodeStopped,
-      isTlsCertInvalid,
-      syncPercentage,
-    } = this.props;
-
-    const showEllipsis = isNodeStopped || (isTlsCertInvalid && !isNodeStopping);
-
-    if (!isConnected) {
-      const headlineClasses = classNames([
-        styles.headline,
-        showEllipsis ? styles.withoutAnimation : null,
-      ]);
-      return (
-        <div className={styles.connecting}>
-          <h1 className={headlineClasses}>
-            {intl.formatMessage(this._getConnectingMessage())}
-          </h1>
-        </div>
-      );
-    }
-
-    if (!isSynced) {
-      return (
-        <div className={styles.syncing}>
-          <h1 className={styles.headline}>
-            {intl.formatMessage(messages.syncing)} {syncPercentage.toFixed(2)}%
-          </h1>
-        </div>
-      );
-    }
-
-    return (
-      <div className={styles.syncing}>
-        <div>
-          <h1 className={styles.headline}>
-            {intl.formatMessage(messages.loadingWalletData)}
-          </h1>
-          <LoadingSpinner />
-        </div>
-      </div>
-    );
-  };
-
   render() {
-    const { intl } = this.context;
     const {
       cardanoNodeState,
       isConnected,
       isSynced,
+      isConnecting,
+      isSyncing,
       hasLoadedCurrentLocale,
       hasLoadedCurrentTheme,
       onReportIssueClick,
@@ -367,6 +177,11 @@ export default class SyncingConnecting extends Component<Props, State> {
       isCheckingSystemTime,
       isNewAppVersionAvailable,
       isNewAppVersionLoaded,
+      hasBeenConnected,
+      isTlsCertInvalid,
+      isNodeStopping,
+      isNodeStopped,
+      syncPercentage,
     } = this.props;
 
     const { connectingTime, syncingTime } = this.state;
@@ -374,29 +189,9 @@ export default class SyncingConnecting extends Component<Props, State> {
     const componentStyles = classNames([
       styles.component,
       hasLoadedCurrentTheme ? null : styles['is-loading-theme'],
-      !isConnected ? styles['is-connecting'] : null,
-      isConnected && !isSynced ? styles['is-syncing'] : null,
+      isConnecting ? styles['is-connecting'] : null,
+      isSyncing ? styles['is-syncing'] : null,
     ]);
-    const daedalusLogoStyles = classNames([
-      styles.daedalusLogo,
-      !isConnected ? styles.connectingLogo : styles.syncingLogo,
-    ]);
-    const currencyLogoStyles = classNames([
-      styles['ada-logo'],
-      !isConnected ? styles.connectingLogo : styles.syncingLogo,
-    ]);
-    const apiLogoStyles = classNames([
-      styles['ada-apiLogo'],
-      !isConnected ? styles.connectingLogo : styles.syncingLogo,
-    ]);
-    const downloadLogsButtonStyles = classNames([
-      styles.downloadLogsButton,
-      !isConnected ? styles.downloadLogsButtonConnecting : null,
-    ]);
-
-    const daedalusLoadingLogo = daedalusLogo;
-    const currencyLoadingLogo = adaLogo;
-    const apiLoadingLogo = cardanoLogo;
 
     const canReportConnectingIssue =
       !isConnected &&
@@ -409,61 +204,39 @@ export default class SyncingConnecting extends Component<Props, State> {
       !isNewAppVersionAvailable &&
       (canReportConnectingIssue || canReportSyncingIssue);
 
-    const buttonClasses = classNames(['primary', styles.reportIssueButton]);
-
-    const isNodeTimeCheckedAndCorrect = isCheckingSystemTime
-      ? undefined
-      : isNodeTimeCorrect;
-
     return (
       <div className={componentStyles}>
         {showReportIssue && (
-          <div className={styles.reportIssue}>
-            <h1 className={styles.reportIssueText}>
-              {!isConnected
-                ? intl.formatMessage(messages.reportConnectingIssueText)
-                : intl.formatMessage(messages.reportSyncingIssueText)}
-            </h1>
-            <Button
-              className={buttonClasses}
-              label={
-                <p>
-                  <SVGInline
-                    svg={linkNewWindow}
-                    className={styles.linkNewWindow}
-                  />
-                  {intl.formatMessage(messages.reportIssueButtonLabel)}
-                </p>
-              }
-              onClick={() =>
-                onReportIssueClick(
-                  intl.formatMessage(messages.reportIssueButtonUrl)
-                )
-              }
-              skin={ButtonSkin}
-            />
-            <br />
-            <button
-              className={downloadLogsButtonStyles}
-              onClick={onDownloadLogs}
-              disabled={disableDownloadLogs}
-            >
-              {intl.formatMessage(messages.reportIssueDownloadLogsLinkLabel)}
-            </button>
-          </div>
+          <ReportIssue
+            isConnected={isConnected}
+            onReportIssueClick={onReportIssueClick}
+            onDownloadLogs={onDownloadLogs}
+            disableDownloadLogs={disableDownloadLogs}
+            isConnecting={isConnecting}
+            isSyncing={isSyncing}
+          />
         )}
-        <div className={styles.logos}>
-          <SVGInline svg={currencyLoadingLogo} className={currencyLogoStyles} />
-          <SVGInline svg={daedalusLoadingLogo} className={daedalusLogoStyles} />
-          <SVGInline svg={apiLoadingLogo} className={apiLogoStyles} />
-        </div>
-        {hasLoadedCurrentLocale ? this._renderLoadingScreen() : null}
+        <LogosDisplay isConnected={isConnected} />
+        <SyncingConnectingTitle
+          cardanoNodeState={cardanoNodeState}
+          hasLoadedCurrentLocale={hasLoadedCurrentLocale}
+          hasBeenConnected={hasBeenConnected}
+          isTlsCertInvalid={isTlsCertInvalid}
+          isConnected={isConnected}
+          isSynced={isSynced}
+          isNodeStopping={isNodeStopping}
+          isNodeStopped={isNodeStopped}
+          syncPercentage={syncPercentage}
+        />
         <StatusIcons
           nodeState={cardanoNodeState}
           isNodeResponding={isNodeResponding}
           isNodeSubscribed={isNodeSubscribed}
-          isNodeTimeCorrect={isNodeTimeCheckedAndCorrect}
+          isNodeTimeCorrect={
+            isCheckingSystemTime ? undefined : isNodeTimeCorrect
+          }
           isNodeSyncing={isNodeSyncing}
+          isCheckingSystemTime={isCheckingSystemTime}
         />
       </div>
     );
